pyproject.toml
[project]
name = "tasklist"
version = "0.1.0"
description = "Ежедневник"
authors = [
    {name = "Sergey Pisarenko",email = "Ser.pisarenko@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "social-auth-app-django (>=5.4.3,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
